requirements.txt
asgiref==3.5.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.1
django-axes==5.13.1
django-cors-headers==3.11.0
django-filter==21.1
django-import-export==2.7.1
django-ipware==3.0.7
djangorestframework==3.13.1
et-xmlfile==1.1.0
firebase-admin==5.2.0
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.5.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
idna==3.3
MarkupPy==1.14
msgpack==1.0.3
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
proto-plus==1.19.9
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0